replace resize with emplace_back when generating candidates
Summary:
Previously, we manipulate the `candidates_` buffer in decoder by using a counter `nCandidates_` and we never free the objects generated in  `candidates_` for efficiency.

But I noticed that if I change everything to simple `emplace_back()` and call `clear()` at the beginning. No speed deterioration is detected.

So I decide to make it simple.

Also, rename some variables. The principle is only use underscore as suffix when the variable is private.

Reviewed By: vineelpratap

Differential Revision: D15607414

fbshipit-source-id: 9d35e531c3258686946902aae47cb70ae3eb7836