check number of parameters when loading fairseq ConvLM
Summary: When we load fairseq convlm to w2l, if the w2l module defined in the arch file has more parameters than the fairseq module defined in the weight file, we will skip those parameters and leave them with the randomly initialized values, and the loading process can still succeed. For example, if we declare a linear layer with bias in w2l arch file but has only weight defined in the fairseq weight file, we won't load anything to the bias. This diff adds check on the total number of parameters.

Reviewed By: awni

Differential Revision: D15718192

fbshipit-source-id: 838848f0eaa304288d8d8d64a3ef8fe59b3368c6