mask padding for lmcrit
Summary: Support masking the padding in the end of each utterance when we compute the LM loss, since in w2l we pad with <eos> during batching, while in fairseq we often train with concatenated sentences with one <eos> symbol in between the sentences. Also support ignoring the first token in the LM loss, since it's always a fixed distribution.

Reviewed By: awni

Differential Revision: D15768421

fbshipit-source-id: 1b74177b0b2ba73b8c33118594a2a9b8942e14a0